requirements.txt
mlflow==2.7.1
jupyterlab==4.0.7
psycopg==3.1.12
psycopg[binary,pool]
pandas==2.0.1
scikit-learn==1.3.1
catboost==1.2.2
scipy==1.11.3
optuna==3.4.0
ipywidgets==8.1.1
seaborn==0.13.0
autofeat==2.1.2
mlxtend==0.23.0
plotly==5.20.0
nbformat==5.10.3
kaleido==0.2.1
boto3==1.34.78
category_encoders
